feat(telex): expand advanced Telex support across platforms
- Enhanced the advanced Telex input method to be exposed on both macOS and Windows, with Linux integration planned for the future.
- Updated configuration format to reflect the inclusion of `telex_advanced` as a valid input method option.
- Improved input method handling in the Windows composer and tray menu, allowing for better user experience and method selection.
- Introduced a new method menu for easier access to input method options in the tray.
- Updated onboarding and settings UI to accommodate the new input method options and provide clearer user guidance.
