+++ platforms/windows/ui/onboarding_method.slint
@@ -1,0 +1,52 @@
+// Input-method step extracted from the onboarding wizard.
+
+import { Palette } from "std-widgets.slint";
+import { Theme } from "theme.slint";
+import { MethodSelector } from "method_selector.slint";
+
+export component OnboardingMethodStep inherits VerticalLayout {
+    in property <string> method;
+    callback pick-method(string);
+
+    alignment: center;
+    spacing: Theme.sp-sm;
+
+    HorizontalLayout {
+        alignment: center;
+        Image {
+            source: @image-url("icons/keyboard.svg");
+            colorize: Theme.accent;
+            width: 42px;
+            height: 42px;
+        }
+    }
+    Text {
+        text: "Chọn kiểu gõ";
+        font-size: 20px;
+        font-weight: 700;
+        horizontal-alignment: center;
+        color: Palette.foreground;
+    }
+    Text {
+        text: "Có thể đổi bất cứ lúc nào trong Cài đặt.";
+        font-size: 14px;
+        horizontal-alignment: center;
+        color: Theme.subtle;
+    }
+    HorizontalLayout {
+        alignment: center;
+        MethodSelector {
+            method: root.method;
+            picked(value) => { root.pick-method(value); }
+        }
+    }
+    Text {
+        text: root.method == "telex_advanced"
+            ? "Full Telex — [→ư, ]→ơ, w đầu từ→ư"
+            : "";
+        visible: self.text != "";
+        font-size: 12px;
+        horizontal-alignment: center;
+        color: Theme.subtle;
+    }
+}
